fix: 초기화시 고객 번호와 업무중 대기열이 초기화 되지 않는 버그 수정

+++ BankManagerUIApp/BankManagerConsoleApp/Models/Banker.swift
@@ -1,12 +1,14 @@
 import Foundation
 
 struct Banker {
-    func work(for customer: Customer, _ delegate: BankerDelegate?) {
+    func work(for customer: Customer, _ delegate: BankerDelegate?, _ isReset: Bool) {
         let task = customer.task
         let processingTime = task.processingTime
         
-        delegate?.addWorkingStackView(self, customer)
-        Thread.sleep(forTimeInterval: processingTime)
-        delegate?.deleteWorkingStackView(self, customer)
+        if !isReset {
+            delegate?.addWorkingStackView(self, customer)
+            Thread.sleep(forTimeInterval: processingTime)
+            delegate?.deleteWorkingStackView(self, customer)
+        }
     }
 }
